+++ batch3dfier/batch3dfierapp.py
@@ -31,24 +31,12 @@ def parse_console_args():
         help="The number of threads to run.",
         default=3,
         type=int)
-    parser.add_argument(
-        "--del-csv",
-        action="store_true",
-        dest="rm",
-        help="Remove CSV files from disk after import (relevant when CSV-BUILDINGS-MULTIPLE)")
-    parser.add_argument(
-        "--keep-csv",
-        action="store_false",
-        dest="rm",
-        help="Keep CSV files from disk after import (relevant when CSV-BUILDINGS-MULTIPLE)")
-    parser.set_defaults(rm=False)
 
     args = parser.parse_args()
     args_in = {}
     args_in['cfg_file'] = os.path.abspath(args.path)
     args_in['cfg_dir'] = os.path.dirname(args_in['cfg_file'])
     args_in['threads'] = args.threads
-    args_in['rm'] = args.rm
 
     return(args_in)
 
@@ -309,46 +297,6 @@ def main():
             cfg['user_schema'],
             views_to_drop=tiles_clipped)
 
-    # If requires, copy the CSV output to postgres
-    if cfg['out_table']:
-        config.create_heights_table(dbase, cfg['out_schema'],
-                                    cfg['out_table'])
-        with dbase.conn:
-            with dbase.conn.cursor() as cur:
-                tbl = ".".join([cfg['out_schema'], cfg['out_table']])
-                for p in out_paths:
-                    # remove trailing commas from the CSV (until #58 is fixed
-                    # in 3dfier)
-                    command = "sed -i 's/,$//' " + p
-                    call(command, shell=True)
-
-                    with open(p, "r") as f_in:
-                        # skip header
-                        next(f_in)
-                        cur.copy_from(f_in, tbl, sep=',')
-
-                    # delete the CSV
-                    if args_in['rm']:
-                        command = "rm" + p
-                        call(command, shell=True)
-                    else:
-                        pass
-        dbase.sendQuery(
-            sql.SQL("""CREATE INDEX IF NOT EXISTS {table}_id_idx
-                    ON {schema}.{table} (id);
-                    """.format(schema=cfg['out_schema'],
-                               table=cfg['out_table'])
-                    )
-        )
-        dbase.sendQuery(
-            sql.SQL("""COMMENT ON TABLE {schema}.{table} IS
-                    'Building heights generated with 3dfier.';
-                    """.format(schema=cfg['out_schema'],
-                               table=cfg['out_table'])
-                    )
-        )
-    else:
-        pass
 
     # Delete temporary config files
     yml_cfg = [
@@ -368,8 +316,7 @@ def main():
     tiles_skipped = set(tiles_skipped)
     logging.info("Total number of tiles processed: %s",
                  str(len(tiles.difference(tiles_skipped))))
-    logging.info("Total number of tiles skipped: %s",
-                 str(len(tiles_skipped)))
+    logging.info("Tiles skipped: %s", tiles_skipped)
 
 
 if __name__ == '__main__':