+++ CHANGELOG.md
@@ -6,11 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [0.7.0] - 2017-12-26
+## [0.7.0] - 2018-01-02
 ### Added
 - Import the contents of CSV-BUILDINGS-MULTIPLE into a postgres table that can be joined on the footprints
 - Log file (just basic info)
 - Python wheel
+- bag3d module that helps to generate a 3D BAG and import the CSVs into postgres
+- `bag3d` console script
 
 ### Changed
 - `batch3dfy` console script -> `batch3dfier`